refactor(Types): move type_class_mapping from Hub to Types
This is going to be needed in upcoming commits to avoid cyclic
imports
